Don't install AppleWWDRCAG3.cer for Big Sur

+++ images/macos/provision/configuration/configure-machine.sh
@@ -21,9 +21,13 @@ sudo "/Library/Application Support/VMware Tools/vmware-resolutionSet" 1176 885
 # Rotate the certificate before expiration to ensure your apps are installed and signed with an active certificate.
 # Confirm that the correct intermediate certificate is installed by verifying the expiration date is set to 2030.
 # sudo security delete-certificate -Z FF6797793A3CD798DC5B2ABEF56F73EDC9F83A64 /Library/Keychains/System.keychain
-curl https://www.apple.com/certificateauthority/AppleWWDRCAG3.cer --output $HOME/AppleWWDRCAG3.cer --silent
-sudo security add-trusted-cert -d -r unspecified -k /Library/Keychains/System.keychain $HOME/AppleWWDRCAG3.cer
-rm $HOME/AppleWWDRCAG3.cer
+# Big Sur asks for user password https://developer.apple.com/forums/thread/671582, we need to make sure this cert is installed on the base image
+if is_Less_BigSur; then
+    curl https://www.apple.com/certificateauthority/AppleWWDRCAG3.cer --output $HOME/AppleWWDRCAG3.cer --silent
+    sudo security add-trusted-cert -d -r unspecified -k /Library/Keychains/System.keychain $HOME/AppleWWDRCAG3.cer
+    rm $HOME/AppleWWDRCAG3.cer
+fi
+
 
 # Create symlink for tests running
 if [ ! -d "/usr/local/bin" ];then
